requirements.txt
cached-property==1.3.1
cairocffi==0.8.0
CairoSVG==2.1.3
cffi==1.11.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-rest-auth==0.9.2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
enum-compat==0.0.2
html5lib==1.0.1
Markdown==2.6.10
pdfrw==0.4
Pillow==5.0.0
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.5.3
Pyphen==0.9.4
pytz==2017.3
six==1.11.0
tinycss2==0.6.1
WeasyPrint==0.42.2
webencodings==0.5.1
